(BSR)[PRO] fix: move parseAndValidateFrenchPhoneNumber
