+++ pro/src/components/pages/Offerers/Offerer/VenueV1/fields/ContactInfosFields/index.js
@@ -1,3 +1,1 @@
 export { default } from './ContactInfosFields'
-export * from './constants'
-export * from './utils'
